[stream] make max message size a u32 (#2569)

+++ p2p/src/authenticated/discovery/config.rs
@@ -44,7 +44,7 @@ pub struct Config<C: Signer> {
     ///
     /// The actual size of the network message will be higher due to overhead from the protocol;
     /// this may include additional metadata, data from the codec, and/or cryptographic signatures.
-    pub max_message_size: usize,
+    pub max_message_size: u32,
 
     /// Message backlog allowed for internal actors.
     ///
@@ -131,7 +131,7 @@ impl<C: Signer> Config<C> {
         listen: SocketAddr,
         dialable: impl Into<Ingress>,
         bootstrappers: Vec<Bootstrapper<C::PublicKey>>,
-        max_message_size: usize,
+        max_message_size: u32,
     ) -> Self {
         Self {
             crypto,
@@ -173,7 +173,7 @@ impl<C: Signer> Config<C> {
         listen: SocketAddr,
         dialable: impl Into<Ingress>,
         bootstrappers: Vec<Bootstrapper<C::PublicKey>>,
-        max_message_size: usize,
+        max_message_size: u32,
     ) -> Self {
         Self {
             crypto,
@@ -208,7 +208,7 @@ impl<C: Signer> Config<C> {
         crypto: C,
         listen: SocketAddr,
         bootstrappers: Vec<Bootstrapper<C::PublicKey>>,
-        max_message_size: usize,
+        max_message_size: u32,
     ) -> Self {
         Self {
             crypto,
