+++ stream/fuzz/fuzz_targets/transport.rs
@@ -7,7 +7,7 @@ use libfuzzer_sys::fuzz_target;
 use std::time::Duration;
 
 static NAMESPACE: &[u8] = b"fuzz_transport";
-const MAX_MESSAGE_SIZE: usize = 64 * 1024; // 64KB buffer
+const MAX_MESSAGE_SIZE: u32 = 64 * 1024; // 64KB buffer
 
 fn fuzz(data: &[u8]) {
     let executor = deterministic::Runner::default();
